All count/list filter options inherit from ShopifyCountFilter, ShopifyListFilter and ShopifyPublishableListFilter. Will close #21.
