+++ ShopifySharp.Tests/ShopifyFulfillmentService Tests/When_listing_fulfillments_with_a_filter.cs
@@ -1,4 +1,5 @@
 ﻿using Machine.Specifications;
+using ShopifySharp.Filters;
 using ShopifySharp.Tests.Test_Data;
 using System.Collections.Generic;
 using System.Linq;
