+++ ShopifySharp.Tests/ShopifyMetaFieldService Tests/When_getting_a_list_of_filtered_metafields.cs
@@ -1,4 +1,5 @@
 using Machine.Specifications;
+using ShopifySharp.Filters;
 using ShopifySharp.Tests.Test_Data;
 using System;
 using System.Collections.Generic;
@@ -39,7 +40,7 @@ namespace ShopifySharp.Tests.MetaField
 
         Because of = () =>
         {
-            MetaFields = Service.ListAsync(Product.Id, "products", new ShopifyMetaFieldFilterOptions() { Namespace = "testing" }).Await().AsTask.Result;
+            MetaFields = Service.ListAsync(Product.Id, "products", new ShopifyMetaFieldFilter() { Namespace = "testing" }).Await().AsTask.Result;
         };
 
         It should_list_metafields_with_specific_namespace = () =>
