+++ ShopifySharp/Filters/ShopifyCollectFilter.cs
@@ -1,13 +1,11 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ShopifySharp
+namespace ShopifySharp.Filters
 {
-    public class ShopifyCollectFilterOptions : ShopifyFilterOptions
+    /// <summary>
+    /// Options for filtering <see cref="ShopifyCollectService.ListAsync(ShopifyCollectFilter)"/> results.
+    /// </summary>
+    public class ShopifyCollectFilter : ShopifyListFilter
     {
         /// <summary>
         /// An optional product id to retrieve. 
@@ -20,6 +18,5 @@ namespace ShopifySharp
         /// </summary>
         [JsonProperty("collection_id")]
         public long? CollectionId { get; set; }
-
     }
 }
