+++ ShopifySharp/ShopifySharp.csproj
@@ -108,17 +108,17 @@
     <Compile Include="Enums\ShopifyAuthorizationScope.cs" />
     <Compile Include="Services\Asset\ShopifyAssetService.cs" />
     <Compile Include="Services\Authorization\ShopifyAuthorizationService.cs" />
-    <Compile Include="Services\Collect\ShopifyCollectFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyCollectFilter.cs" />
     <Compile Include="Services\Collect\ShopifyCollectService.cs" />
-    <Compile Include="Services\CustomCollection\ShopifyCustomCollectionFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyCustomCollectionFilter.cs" />
     <Compile Include="Services\CustomCollection\ShopifyCustomCollectionService.cs" />
     <Compile Include="Services\Fulfillment\ShopifyFulfillmentService.cs" />
-    <Compile Include="Services\MetaField\ShopifyMetaFieldFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyMetaFieldFilter.cs" />
     <Compile Include="Services\MetaField\ShopifyMetaFieldService.cs" />
     <Compile Include="Services\Page\ShopifyPageService.cs" />
-    <Compile Include="Services\Page\ShopifyPageFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyPageFilter.cs" />
     <Compile Include="Services\Page\ShopifyPageCreateOptions.cs" />
-    <Compile Include="Services\Product\ShopifyProductFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyProductFilter.cs" />
     <Compile Include="Services\Product\ShopifyProductCreateOptions.cs" />
     <Compile Include="Services\Product\ShopifyProductService.cs" />
     <Compile Include="Services\RecurringCharge\ShopifyRecurringChargeService.cs" />
@@ -126,17 +126,16 @@
     <Compile Include="Services\Customer\ShopifyCustomerService.cs" />
     <Compile Include="Enums\ShopifyOrderCancelReason.cs" />
     <Compile Include="Services\Order\ShopifyOrderCreateOptions.cs" />
-    <Compile Include="Services\Order\ShopifyOrderFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyOrderFilter.cs" />
     <Compile Include="Services\Order\ShopifyOrderService.cs" />
     <Compile Include="Enums\ShopifyProcessingMethod.cs" />
-    <Compile Include="Services\Redirect\ShopifyRedirectFilterOptions.cs" />
-    <Compile Include="Services\ScriptTag\ShopifyScriptTagListOptions.cs" />
+    <Compile Include="Filters\ShopifyRedirectFilter.cs" />
+    <Compile Include="Filters\ShopifyScriptTagFilter.cs" />
     <Compile Include="Services\ScriptTag\ShopifyScriptTagService.cs" />
-    <Compile Include="Services\ShopifyCountFilter.cs" />
-    <Compile Include="Services\ShopifyFilterOptions.cs" />
-    <Compile Include="Services\ShopifyListFilter.cs" />
-    <Compile Include="Services\ShopifyListOptions.cs" />
+    <Compile Include="Filters\ShopifyCountFilter.cs" />
+    <Compile Include="Filters\ShopifyListFilter.cs" />
     <Compile Include="Services\Redirect\ShopifyRedirectService.cs" />
+    <Compile Include="Filters\ShopifyPublishableListFilter.cs" />
     <Compile Include="Services\ShopifyService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\Shop\ShopifyShopService.cs" />
@@ -145,7 +144,7 @@
     <Compile Include="Services\Charge\ShopifyChargeService.cs" />
     <Compile Include="Services\Theme\ShopifyThemeService.cs" />
     <Compile Include="Services\Transaction\ShopifyTransactionService.cs" />
-    <Compile Include="Services\Webhook\ShopifyWebhookFilterOptions.cs" />
+    <Compile Include="Filters\ShopifyWebhookFilter.cs" />
     <Compile Include="Services\Webhook\ShopifyWebhookService.cs" />
   </ItemGroup>
   <ItemGroup>
